(BOUNTY) Psuedo-Respiration Quirk (#4241)
## About The Pull Request
Adds a negative hemophage specific quirk that gives hemophages the
ability to breathe. That is all this does, they will not be susceptible
to diseases or lose any of their other race quirks.
## Why It's Good For The Game
It allows for more customization of your character as a hemophage, and
allows for players to 'hide' their condition better.
## Proof Of Testing
It worked fully on my localhost when adding and removing the quirk. 
## Changelog
:cl:
add: added the pseudo-respiration quirk, a negative quirk that allows
hemophages to breathe.
/:cl: