Add volume slider pref for emote sounds (#4301)
## About The Pull Request

Add a new volume pref in game options that controls how loud emotes that
have sound effects are.
Note: This PR was made to fulfill a bounty by @NanoCats
## Why It's Good For The Game

Quality of life improvement, we have many emotes that play sounds here
and not everyone wants to hear them. Also there's that one scream that's
insanely fucking loud

## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>
Note: the emote cooldown was disabled in this video for demonstration
purposes


https://github.com/user-attachments/assets/1e465ad1-6be9-4113-a564-72ace0c78118

</details>

## Changelog
:cl:
qol: added a volume slider for emote sounds to game options
/:cl: