renames the guilmon + renamon parts to toraxen + cyvian by headmin request (#4295)
## About The Pull Request
One of our server rules is that you can't play species from other canons
1:1 so it's a bit misleading to have parts named after them. Headmin
wanted them to be renamed. Will probably do the same for eevee+scolipede
when they come up with names for it. Leaving the datum names the same
and the like so it's still clear what it's meant to be if you decide to
codedive.

## Why It's Good For The Game
code follows standards set by server rules this is a good thing

## Proof Of Testing
yeah it works

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl:
code: Renames the guilmon + renamon parts to toraxen + cyvian
respectively.
/:cl: