Logging for bot sapience (#92302)
## About The Pull Request

This PR adds logging for people turning bot sapience on and off
This is currently logged to silicon logs, even though the mob is not a
silicon, just because I figured that's where an admin would look for
information related to robots. If you think that is stupid I can put it
somewhere else.

## Why It's Good For The Game

This is an action that can adds ghosts (and meta knowledge) into the
round which didn't leave any kind of paper trail, so it should probably
have one
Ejecting an existing player from bot control was already logged and I
didn't remove it

## Changelog

:cl:
admin: adds logging for turning bot sapience on and off
/:cl: