Gives Changelings Zombies the FAKEDEATH trait (#4243)
## About The Pull Request

Gives changeling zombies the FAKEDEATH trait, which should make them
show up as dead on HUDs, like with High Functioning Zombies.

This quite possibly has knock-on effects due to the use of FAKEDEATH,
but that will be seen.
## Why It's Good For The Game

dead :)

## Proof Of Testing

to be tested

## Changelog
:cl:
add: Changeling zombies now show up as dead on health HUDs.
/:cl:

Co-authored-by: Arturlang <24881678+Arturlang@users.noreply.github.com>