Adds Shrinkage and fixes some display bugs when examining Johnsons. (#4217)
Copy of #3344

## About The Pull Request

Fixes some various display bugs when examining Johnsons. The width of a
Johnson is now measured in diameter instead of circumference.
Having a cold Johnson will cause it to shrink.

## Why It's Good For The Game

We're an atmos sim. We should have atmos sim mechanics.

## Proof Of Testing

Tested in the previous PR ages ago but it should still work.

## Changelog

:cl: BurgerBB
add: Having a cold Johnson will cause it to shrink.
fix: Fixes some various display bugs when examining Johnsons. The width
of a Johnson is now measured in diameter instead of circumference.
/:cl: