Makes Nakamura Modsuits Not Be An Import Company Anymore (#4291)
## About The Pull Request
Nakamura Modsuits will no longer be an import category on the cargo
consoles. Yknow? The one that sells modsuit upgrade parts?

Single cores moved to goodies, magnetic deployable moved to goodies.

Also normalizes the price for mod cores between single and three-pack
## Why It's Good For The Game
Y'know with a focus on Not Having People Avoid Interaction With Jobs For
Little To No Reason (addition of stay in your lane rule) I got inspired
to get rid of the nakamura modsuits category altogether on company
imports due to mechanically supporting such behavior.

1. A lot of the orderable parts on there are both really cheap, and
several nodes deep into modsuit tech. For example, you can order the EMP
protection module for a measly 50 credits. There are similar costs for
jetpack, magboots, sphere, atrocinator, and so forth.

2. This ties in to my second point, when things are this readily
available, why would people ask the job with the dedicated fabricator
for this to do said job? Sure, you could go to the roboticist who may or
may not be busy to ask them to print modules, but why would you do that
when you can access the things you need with your roundstart money. Sure
as hell also eliminates the wait for research.

3. The category was made before modsuits were accessible as they are
now. Proteans exist. Entombed exists. These are both species/quirks
which have modsuits on 100% of the time, and it's not uncommon for these
people to do what I've outlined in point number two shiftstart. It turns
what is effectively a downside on proteans (regular storage module
instead of expanded) into a nonfactor when it can be ordered the instant
they spawn in as opposed to - well - having player interaction occur at
all.

Yes, the costs could be raised, but the economy on bubberstation is
already in a nonfunctional state. If you spend 5-10 minutes fishing, or
god forbid, grab a few sheets of plasma and sell those (of which there
will typically be plenty for what you're doing not to be an impact not
too long into a round) you could afford them even with a major price
increase. I do not believe it would stop the problematic behavior in
question.

It's heartbreaking to see that the roboticists (let alone the rest of
the crew) is better off ordering everything off space amazon even if
science is filled with Cuban Fucking Pete and the Ghost of Einstein
themselves.

P.S. I'm making a seperate PR to give proteans expanded roundstart as a
compensation to this. If entombed gets it, why not them?
## Proof Of Testing
I ran it on local and the section is no longer there
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
del: Removes Nakamura Modsuits from the list of company imports
selections. Some of the items have been moved to the goodies tab.
balance: Raises the price of single mod cores (now in goodies) to 200
credits a piece, to match the price of a three-pack as opposed to being
half the price if you buy them individually.
/:cl: