Bubberstation gun replacements part 1: Murphy service pistol [part of skyrat removal PR] (#3949)
## About The Pull Request
Feedback here:
https://discord.com/channels/1059199070016655462/1381309027849474169
This PR is a complimentary PR that fills in the widest gaps left by the
now-removed arrangement of guns. The focus is on making security still
have a very reliable sidearm that is interesting and provides relevance
even in the later game. It is also perfectly usable as a regular weapon
without all the fluff.

This PR is one of 5 total, and with being asked to atomize my things, it
will be the first one to try out for testmerge.

# This PR only adds the murphy, rest is data on all plans and guns

## 'Murphy' security pistol
``This 9 mm monster was developed during the very first body-modding
craze by NanoTrasen, built with every what-if in mind, this timeless
brick is near-incapable of failure. 60% more bullet-per-bullet ammo may
not have helped it's killing power, but it sure helped this gun hit like
an actual brick when used as a bludgeon. Not only that, but it's
'anti-lawsuit' heavy trigger design allows for safe spinning, if one can
handle it and not hit anyone with the magazine flying out the
handle...``
- 20 damage shots
- 10 bullet magazines
- baton-equivilent pistol whip damage
- increased throw force
- press Z when the gun is dry to reload your gun over (2-3 seconds,
DRAFT, as I said). This can be done on the move and you do not need your
other hand empty to do this.
- toggleable setting, off by default, ejects your magazines for a tiny
bit of extra damage.
- researchable magazines that pack only 8 rounds in return for dealing
[30-40, up to adjustment] damage. [knockdown is being discussed]

## Upcoming: Personality Core Eguns, The SC-3
Aren't you tired of old eguns? The blandness of them? Their low amount
of potential upgrades?
Well fear not, as with every arms race, Allstar has made developments of
their own.
*lightly* copying the homework off their more morally corrupt
competitors, personality modules.

So how's it compare to an egun? 
- Radial menu is a bit clunkier to use
- Comes with a free combat knife as a module
- Will loudly proclaim your location when emp'd due to a personaltiy
module
- Is practically the same combat-wise as the egun if you were to
micromanage a combat knife alongside it. It's a third-degree, but was
highly requested, and not only! (I say equivilent to combat knife
because I am about to release an ERT buff PR)

**As you can impact what the final modules are with this poll**
https://forms.gle/d3Fst6ZLWYTb8mDo7


--## Plasma Gun [the most wip of them all]
- Magazine fed bulky weapon
- We hear concerns about vampires being too strong, so this weapon's
shots damage scales depending on the amount of shots a target has
received from this particular gun. No, 5 officers with the same gun will
not be more effective than one.
- Due to it's high power (if you can mind the damage rampup), it's also
highly expensive and volatile. You shouldn't let one of those armaments
explode, lest it's core bursts. (Gibtonite effect upon being caught in a
heavy explosion, after some time).
--

This weapon is not needed for the game balance, instead, focus is
shifted to rebalancing the Kiboko to be a usable gun. The aim will be to
have it to take on a more supportive role, being effective at applying
firestacks and providing smoke if need be.

The kiboko will become a tool that will require itself to be roundstart
only, or at least dep-order only as to prevent it from being a recurring
threat if dealt with once already.

- Buff damage from 20 -> 30 on hit
- Buff projectile velocity
- Remove practise and shrapnel rounds
- Remove delay between hitting target and exploding (if on the same
tile)
- Grenades will now explode even if they score a direct hit
- Less deadly fire grenades (up to 100+ damage per one currently if met
on an unarmored target in close proximity, will likely make it less
damaging but more consistent)


Planned addition:
       Breacher rounds:
The kiboko's breacher rounds are inspired by the strange delay they
initially had and how it felt reminiscent of R6's Ash. Frangible slugs
are not often used as it takes away ammo from a main weapon, they are
also clunkier and require several rounds. They have their place but are
overall quite messy.

What breacher rounds will do is cause a delayed (mostly cosmetic)
explosion after striking a target object, structure or person. The delay
will be larger if the integrity of the object hit is large, scaling.
This isn't the end of the world as it will still be the quickest or at
least easiest way to tear down walls and airlocks. This also provides
some time to take out any shields, weapons and prime grenades while
waiting on the charge to drill deeper and have the fuse detonate.
   
     
## Why It's Good For The Game
(will update this, currently my exams are approaching very quickly, so
please be understanding for needing yet more time. Work speed will pick
up drastically once that is over. )

New security get a gun that is easy and convenient to use.
The gun is not meant to be your main tool of dealing with threats such
as traitor, rather your backup plan. As such, it benefits from okay-ish
stats for melee combat once you run out of everything.
It shakes up the belt-storage meta, making magazines more enticing of a
choice considering the potential of unloading very consistent fire if
one is to use that room.
For threats that are more slippery or plentiful such as carps or spiders
this gun's aforementioned consistency crowns over other ones.
Lastly, in those scenarios it helps you keep your shield up and keep
dishing out damage. Turn off your brain for once with the Murphy.

Highly requested bit of fun gun customization, will come with multiple
personality matrices.

~~
Helps security and crew with dealing with extremely defensive and bulky
threats. Be it a nukie, mech or vampire, if the user maintains healthy
social distancing then the tables will turn quickly, forcing antagonists
to reconsider their approach and introduces a gun that while possible to
use personally, becomes much better with team-oriented protective
playstyles.

~~ I have discussed this heavily with my team and they have reassured me
that the current threats are not as bad as it would have initially
seemed. The tools to deal with that already exist.

The kiboko's rounds can be outran which makes it only a good breaching
weapon, as a breaching weapon it is outclassed by almost anything else
as anyone can block the grenade with their body for a measely 20 damage
pre-armour, and then there's the fact it's shrapnel rounds cause weeping
avulsions near-guaranteedly if they hit. It's a mess.

Again, disclaimer, the values are not final, will be likely changed,
aspects such as how the guns work are being discussed and the visual
aspects are going to be handled last.

The necessity of more is being discussed as 5 concepts of guns were
scrapped, 2 of which mid-development.

## Proof Of Testing
To be done, you can find discussion pertaining this in the skyrat gun
removal thread.


## Changelog

:cl:
add: Added a security sidearm - the Murphy service pistol. Spin to
reload with ammo from your belt slot, toggle launch-ejecting mags (and
special mags) for a little (or a lot) extra damage. Show the last person
to have laughed at your gun why your 'brick' is called a brick by
pistolwhipping them (baton equivalent damage on hit, stronger throw).
add: 9mm security caliber for the murphy.
/:cl:

---------

Co-authored-by: Roxy <75404941+TealSeer@users.noreply.github.com>