Lets cyborgs filter medical combitool (#4322)
## About The Pull Request

Quick fix to let cyborgs add filters to the combitool with CTRL click.
Ideally, the cyborg click code would be fixed for this. But I am not
skilled enough at coding or modularization to do that
## Why It's Good For The Game

Cyborgs should have access to the same functions of the tools they are
given that carbons do

## Proof Of Testing

Compiled, CTRL clicking as cyborg opened whitelisting menu

## Changelog
:cl:
fix: Cyborgs can now add filters to medical Combitools with CTRL click
/:cl:

Co-authored-by: KeptheKobold <>