Marked One re-add (#4313)
## About The Pull Request

Basically: we're re-adding the marked one, nerfing the stats on it's
armor into the ground- (and technically below ground too!) so it's more
in line with the design goals of the maintainers. The concept is simple;
by directly making it so you'll take *more* damage by wearing the armor,
it would be monumentally idiotic to go out and try to "frag out for
valids" in armor which renders you more vulnerable to conventional
weapons. You'd be better off using a basic survival suit than this set
of armor.

## Why It's Good For The Game

re-adds a long, long lost fight from a while ago- essentially putting
more content back in the game and a optional "final boss" for the
experienced, or not-so-experienced shaft miner to try their hand at.

It also addresses a long standing gripe- that shaft miners take their
fancy new toys and steam roll other players / antagonists, or go
proto-murderboning in their suits of indestructible fuck-you armor. This
armor *literally* is worse than wearing nothing; thus severely limiting
it's practicality to lavaland only and it's related ecosystem.

## Proof Of Testing

Going to comment my proof of concept because github isn't working right
for me, standbye.

## Changelog

:cl:

balance: heavily nerfs the berserker armor's stats to make it more in
line with a SEVA suit
balance: drastically increases the difficulty of the marked one fight,
giving his attacks armor penetration
balance: re-adds the marked one ruin and bossfight now that he's no
longer lobotomized.

/:cl: