Replaces the Wizard Dice event with a round-start wizard role (#4232)
## About The Pull Request
Comments out the Die of Fate event and replaces it with a round-start
wizard antagonist that happens ***VERY*** rarely. Said role has a weight
of 2 and requires a minimum of 35 players to be readied up for it to
even roll. The die of fate that spawns in trash piles has also been
replaced with the Lavaland variant.

_The die of fate event is still spawnable by admins._
## Why It's Good For The Game
This is going to be a controversial one, so let me run you guys through
my thought process. Currently, as it stands, the wizard die event causes
players to turn into mindless dice-hunting zombies that will employ
various LRP tactics to get the die, ranging from rushing to cargo to get
a Ouija board to simply tiding into places where they see the die. This
event also runs quite frequently, leading to frustration among players
in the departments that get tided into. Burger's original reason for
adding the die of fate event was to "add wizard naturally into the game
without a ghost role for players to idle for". This will prevent players
from idling for it as it happens at round-start. This is the fairest
compromise I could come up with for players to be able to roll wizard
without the tiding and general tomfoolery that happens because of the
dice.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

<img width="668" height="196" alt="image"
src="https://github.com/user-attachments/assets/46dec2f9-a4ea-4d8e-9f85-e9e90e432ac0"
/>


</details>

## Changelog
:cl:
add: added a round-start wizard event
del: removed the wizard die mid-round event
balance: replaced the maintenance loot wizard die with the proper die of
fate
/:cl: