Makes Heretic more common but not overwhelmingly so (#4408)
## About The Pull Request
Updates Heretic's storyteller weight to 5 and their minimum population
required to 30. This puts them on-par with Bloodsuckers as a rare but
not "Literally once every 50 rounds" rare antag.

Also my power went out twice in the making of this PR which is an ill
omen and a half.
## Why It's Good For The Game

Heretics cool. Too many Heretics not cool. Make Heretics slightly more
common. But not too common. Leaves the cap untouched so at most you will
only see two Heretics at once rather than the four of the test period.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
balance: Makes Heretics slightly more common (but less so than the test
period)
/:cl: