Allow command to have missing limbs on roundstart. (#4311)
## About The Pull Request

As the code currently is, if any missing limb is selected, command is
locked. This PR adjusts the logic so that command is locked only if both
arms are missing.
## Why It's Good For The Game

Disabled captains and other heads of staff offer more character creation
and roleplay opportunities. This change removes such logic discrepancies
such as a paraplegic captain being fine bug a legless captain being
disallowed. Crew with one hand are still very capable and do not need to
be blocked from command roles.
## Proof Of Testing

Code compiles with 0 errors, 0 warnings. Tested locally and working as
expected.
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
balance: Heads of staff are free to be missing as many roundstart limbs
as they want, as long as they have at least one hand.
/:cl: