New Shadekin Tail + Ears (#4325)
## About The Pull Request

This adds a lot of new ears, and a new 64x32 tail sprite. To facilitate
the addition of this sprite, I have also relabeled one of the 64x32
sprite folders to reduce clutter. This has not altered the functionality
of the sprites contained within. The sprites are ported from the Goob
Station codebase on SS14, which are further ported from a place called
"EE".

## Why It's Good For The Game

Customization!

## Proof Of Testing



https://github.com/user-attachments/assets/6edbb375-c6b7-4f27-9172-4fca1078144a


## Changelog
:cl:
add: Added new Shadekin Ear options
add: Added a new Shadekin tail
/:cl:

Co-authored-by: Roxy <75404941+TealSeer@users.noreply.github.com>