Fixes MOST /tg/ underclothing items not having digitigrade sprites (#4416)
## About The Pull Request
Did you know that we had code in place and the icons in the code for
digitigrade variants of 99% of the /tg/ clothing items? Apparently we
have, for ***3 years***, it just didn't have an update flag for the
digitigrade variants to be recognized. This fixes that.
## Why It's Good For The Game
Bugs are bad, especially ones that have existed for three years and have
prevented customization.
## Proof Of Testing
All of the cargo undersuits worked and none had funky icons, so I would
assume this is the same for all of them. I'd still recommend this gets
test merged first so I can identify and fix any oddities.
## Changelog
:cl:
fix: fixed a lot of digitigrade variations of undersuits not showing up
/:cl: