Moving on: Updates human lore (#4428)
## About The Pull Request
This PR re-does the human lore listed on the server to be consistent
with all recent updates.

It has yet to be reviewed because it's 3:40am, but the PR is fully
functional. It also lacks the section on the Pan-Slavic Commonwealth
which I will add later. Will mark as ready for review when it's ready
for TM.

## Why It's Good For The Game
Keeps humans consistent with all the other lore updates we've had, and
improves humans' place in the universe.

Humans are space orcs!

## Proof Of Testing
<img width="915" height="540" alt="image"
src="https://github.com/user-attachments/assets/20aa9101-71a5-464a-83aa-4f71dda78135"
/>

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender, Bangle (Lore)
add: Humans have received some substantial updates to their lore,
including the beginnings of a timeline. Check it out!
/:cl: