Storyteller Playercount Calculation Overhaul (#3988)
## About The Pull Request

Significantly overhauls how players are counted when running events. It
now considers where the player actually is, what (who) they are doing,
and if they are actually a crewmember.

## Why It's Good For The Game

Previously, events would use a proc that'd get every single active
living player, regardless if they were crew or not or where they were in
the map. This includes people on the interlink, ashwalkers visiting the
station for """roleplay""", or player-controlled medibots roaming around
the station looking for people to say "fuck you" to.

Now, it uses a calculation that actually counts players who are engaging
in the game so we don't get situations where blob ran because
technically there are 20 players playing (5 are on interlink, 3 are in
dorms, 2 are ashwalkers, and 10 are actually on the station).

A player is considered active if they are:
- In the living mob list.
- They are not AFK.
- They signed up as a job, and not an off-station role.
- On a station z-level.
- Not in dorms.

## Proof Of Testing

Active check works. I can't really test anything else because I don't
have friends.

## Changelog

:cl: BurgerBB
balance: Significantly overhauls how players are counted when running
events. It now considers where the player actually is, what (who) they
are doing, and if they are actually a crewmember.
/:cl: