Fixes Upload Turrets on Nebulastation; Adds External Secuirty Camera (#92580)
## About The Pull Request

Makes it so that Turrets in the AI Upload can actually fire into the
center of the room where the AI and Cyborg Upload Terminals are. Adds an
External Camera to the Vault overlooking the AI Upload.

## Why It's Good For The Game

Hardens the Upload against Intrusion, primarily from the inside, but
also adds some external detection to that area of space so that any
surprises from outside have some degree of forewarning. Turret Placement
modification should let the AI be at less of a mercy to would-be
intruders.

## Changelog

:cl:
map: Adjusts Turret Placement in the AI Upload on Nebulastation
map: Adds an External Security Camera to the Starboard Vault Wall
overlooking the AI Upload on Nebulastation
/:cl: