The Drunken Nanite (#4450)
## About The Pull Request

I went through every drink and checked whether or not it had a metal in
it, and if it did, I made it so a protean can drink it. A nice bit of
flavour that goes with the fact that they use metals as a resource. You
add metals to a drink, a protean gets effected by them, and if a metals
infused with alcohol, well it seems fitting that it gets them a little
tipsy~ I also gave them the synth drinks but im not too tied to that
idea, but they are still synthetic so I threw em in.

## Why It's Good For The Game

At the moment, they have no drinks that effects them. This gives some
roleplay and flavour to them.

## Proof Of Testing

I also checked to ensure it doesn't allow Synths to drink them, it did
not.

https://i.imgur.com/CXJlS4L.png

</details>

## Changelog

:cl:
add: Proteans can now process Synthetic Drinks and Drinks with Metals
in.

/:cl: