[Config Change] Disables Vox_Primalis as a roundstart species (#4471)
## About The Pull Request
Title
## Why It's Good For The Game
Vox_Primalis is un-played, and un-maintained. There's 2 sets of vox
species as of right now, the one with the old sprite, and the one with
the fancier sprite. All of our vox players swear by the old sprite ones,
not that there's many vox players even keeping that in mind though.
Having 2 seperate voxes additionally, is Hella Fucking Confusing.

I'll be working on the full removal of them in the meanwhile but that
might take a while because uhh life = busy
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
config: disabled vox_primalis as a roundstart species
/:cl: