Restores traitor to 20tc instead of 25tc because i died to it once (#4402)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
puts traitor back at 20tc

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Traitor, on bubber/skyrat, has had 25tc for quite some time. In part
this was due to our lack of progtot (which our upstream has since gotten
rid of and restored to the flat 20), and in other parts due to traitor
being quite weak compared to security at its prime. But truth be told,
5tc/25% is a LOT more than it seems.

This inflated amount of currency allows for quite a few combos which are
less than fun to fight against. The many scarp + (something else) combos
are some of the major examples. Whilst it's a major problem player in
the economy, it is not worth being the sole balancing factor for
cost.Discounts DO impact this, if you get lucky, you can snowball even
further due to this, however, it's a mere symptom of the problem, as
opposed to the root cause. The raised cost cap does allow for more
nonsense. For example, if you get an esword + ebow + noslips, that's
18tc. Under normal, no-discount scenarios, this would prevent you from
also getting armor. However, with the raised amount of TC we have, this
allows for you also get an infiltrator mod. A full 50% damage reduction.
This SIGNIFICANTLY increases your survivability.

While yes, in theory, you could balance the costs of EVERY single item
to work with the raised limit, this would be quite the effort to
actively maintain, especially with how our upstream does regularly
change/add stuff that is within the realm of traitor's gear. This change
will prevent some of the more ridiculous goings-on, traitor has plenty
good stuff even without that extra 5.

Also skyrat guns are confirmed to be going when maintainers review the
PR so both sides lose power.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Sets traitor and loneinf's TC count to 20 as opposed to the 25
they had before.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->