Lore language rewrites (#4528)
## About The Pull Request
This PR simplifies a lot of the language text, making them take up less
space in the languages menu. It also brings many of the language
descriptions in line with what our current lore for them is, and
prevents overlap.

## Why It's Good For The Game
Gets information across more readily and easily, creating a better
experience overall

## Proof Of Testing
It compiles

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
qol: Various languages have been rewritten or simplified in the language
menu
/:cl: