Adds the hand labeler to the loadout (#4520)
## About The Pull Request

Title.

It's in the other section.
## Why It's Good For The Game

Hand laborers have many uses and are so so nice for organization, but
are rarely used because you need an autolathe and it's generally a bit
of a hassle for something so minor?

Also, I just think it'd be neat and I'd probably use it a lot.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
add: Loadout other section now has a hand labeler
/:cl: