Fixes a runtime in Job Intern code (#4543)
## About The Pull Request
This PR changes the code for job interns to prevent a runtime related to
jobs that have no real /datum/job as a department head (such as NTC and
Blueshield.) Now, if the job doesn't have a department head, it'll just
move on instead of runtiming.

## Why It's Good For The Game
Fixes an error, no player facing changes

## Proof Of Testing
I sadly cannot test this. It compiles

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
fix: A runtime in Job Intern code for jobs lacking a department head to
reference
/:cl: