De-emphasizes beaks in Teshari lore as we don't have any beak snouts (#4571)
## About The Pull Request
This PR does two things:
One, helps with clarity by cleaning up some of the passage, particularly
de-emphasizing Teshari beaks, as we don't currently have beak snouts
available for Teshari.
Two, removes mentions of Penelope 3, anonymizing it until we have a
replacement name. This is, as I understand, the only part of Minius'
Teshari lore that we've left in, and he wasn't pleased with that. We
shouldn't have any issues with licensing related to lore in the future.

## Why It's Good For The Game
Helps with clarity, doesn't recommend a type of snout we don't currently
have made for Teshari

## Proof Of Testing
String edits, it compiles

## Changelog

:cl: ReturnToZender
del: We no longer use the name Penelope 3 in the Teshari lore blurb, and
the emphasis for beaks has been flipped. Sirisai is exiting an ice age,
muzzles are more common.
/:cl: