Effectively disables monkey and mushroom languages (#4525)
## About The Pull Request
This PR disables two of the worse languages present on the repo, both
are related primarily to seasonal events that only come around
periodically. Mushroomspeak, and Monkey.

Monkey remains available to synthetic and human-adjacent species, as
well as anthromorph.
## Why It's Good For The Game
The extra languages bloat the languages menu, and these two aren't
particularly used by anyone, so I felt it appropriate to disable it
during the making of #4523.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl: ReturnToZender
del: Mushroomspeak and Monkey are limited to their respective species.
/:cl: