New Quirk: Oviparity (Egg production) (#4577)
## About The Pull Request

### Introducing... A NEW QUIRK! Oviparity!
**That's right. Egg production is now Real™!!** Bringing a simple,
interactive egg production system that makes use of certain reagents in
order to produce, well, eggs!

However that's not all...
### WE'RE MAKING THE MOTHER OF ALL OMELETTES HERE, JACK! CAN'T FRET OVER
EVERY EGG!
But with this quirk, you WILL, in fact, be able to (and be fully
justified in doing so to your partner) fret over every egg. With a
dynamically scaling move speed slowdown, you'll FEEL the weight of all
the eggs you've accumulated! **_Every. Single. Egg._** And it can get
_quite_ punishing, so make sure you're clocked out or not planning on
doing anything (or going anywhere) important beforehand if you're going
all in!

<img width="557" height="183" alt="oviparity3"
src="https://github.com/user-attachments/assets/ab9c8517-d806-433f-b700-944e1f87e163"
/>

These also come with unique and special texts to indicate when you're
being Productive™, when you've reached certain thresholds, and even when
you've gone back under those thresholds! (Currently set to every 20% of
the max).

<img width="868" height="459" alt="oviparity1"
src="https://github.com/user-attachments/assets/1b1de88c-376e-46ae-a93e-5b1445716b8e"
/>
This quirk makes use of reagents such as cum, crocin, hexacrocin, and
teshari mutation toxin to induce the production of eggs in your body,
storing them all in a counter that you can keep track of via the
"Produce an Egg" button.

<img width="192" height="175" alt="oviparity2"
src="https://github.com/user-attachments/assets/40a9e25b-1270-495c-a9a2-7b83db4bdba6"
/>

Oh right, you can also lay the eggs anytime you wish, make sure you've
got lots of free time before doing so though...!
### Built with Anonymity in mind!
This quirk was designed with many concerns considered and measures
applied and taken to ensure that those that prefer not to interact with
it, can easily choose not to and avoid it! Eggs laid are automatically
put in hand if they are empty, there are NO auto-emotes whatsoever, and
as it is a manually controllable ability, you will always have the
option to avoid using this in crowded areas, making it easy and clear to
tell when someone is intentionally abusing their egg-laying
capabilities!

## Why It's Good For The Game
There are absolutely zero features, or any other interactive system for
eggs and carbons in the game currently and with this quirk we aim to fix
that with what is hopefully just the start of a neat idea. New fun
interactable stuff like this is always good right?!
Also it finally makes use of cum as a reagent, something that currently
has no real significant mechanical interaction with stuff in-game, which
we thought felt rather strange.

## Proof Of Testing

If we didn't test this I don't think we could confidently or in good
conscious PR this feature ever or at all. As our third ever time coding
something for SS13 in byond/dm code, this was certainly a learning
journey and we took every measure to test everything we could.
On that note, here's some hopefully detailed videos.
<details>

Basic Functionality: 

https://github.com/user-attachments/assets/5766b650-45ad-4673-be54-25b114c70cd3

Threshold Showcase (every 20% of the max, a special text will pop up to
indicate you're getting closer to being completely full):

https://github.com/user-attachments/assets/a078914a-4e4b-4155-8588-b6f5f710ac4d

Proof of concept that this functions even with Teshari Mutation Toxin,
and between species swaps:

https://github.com/user-attachments/assets/a1f3c997-0f11-462d-869b-e2d631ebc02f

</details>

## Changelog
:cl: Astrum Lernaean
add: Added a new Quirk: Oviparity (Egg Production). Gives the ability to
lay eggs (similar to the abductor organ), and allows for reagents such
as cum, crocin, hexacrocin, and teshari mutation toxin to induce egg
production.
tweak: Swapped out Smooth-Headed's icon for a more fitting one, and gave
Oviparity the egg icon. Egg belongs on egg quirk.

/:cl: